Further reductions
